refactor: flatten deployment access tab structure
