+++ web/features/deployments/detail/settings-tab/__tests__/channels-section.spec.tsx
@@ -1,7 +1,7 @@
 import type { AccessChannels, AccessEndpoint } from '@dify/contracts/enterprise/types.gen'
 import { render, screen } from '@testing-library/react'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
-import { deploymentRouteAppInstanceIdAtom } from '../../../../route-state'
+import { deploymentRouteAppInstanceIdAtom } from '../../../route-state'
 import { AccessChannelsSection } from '../channels-section'
 import { accessSettingsQueryAtom } from '../state'
