+++ web/features/deployments/detail/settings-tab/channels-section.tsx
@@ -8,9 +8,9 @@ import { useAtomValue } from 'jotai'
 import { useTranslation } from 'react-i18next'
 import { SkeletonRectangle, SkeletonRow } from '@/app/components/base/skeleton'
 import { consoleQuery } from '@/service/client'
-import { DeploymentEmptyState, DeploymentNoticeState, DeploymentStateMessage } from '../../../components/empty-state'
-import { deploymentRouteAppInstanceIdAtom } from '../../../route-state'
-import { Section } from '../../components/section'
+import { DeploymentEmptyState, DeploymentNoticeState, DeploymentStateMessage } from '../../components/empty-state'
+import { deploymentRouteAppInstanceIdAtom } from '../../route-state'
+import { Section } from '../components/section'
 import { CopyPill, EndpointRow } from './common'
 import { accessSettingsQueryAtom } from './state'
 import { getUrlOrigin } from './url'
